# docs/swagger.yaml
basePath: /api/v1
definitions:
  request.DebtorPayInstallmentLinePayload:
    properties:
      installment_line_id:
        type: string
    required:
    - installment_line_id
    type: object
  request.DebtorTransactionPayload:
    properties:
      asset_name:
        maxLength: 255
        minLength: 1
        type: string
      otr:
        type: string
      tenor_limit_id:
        type: string
    required:
    - asset_name
    - otr
    - tenor_limit_id
    type: object
  request.LoginPayload:
    properties:
      email:
        type: string
      password:
        maxLength: 25
        minLength: 5
        type: string
    required:
    - email
    - password
    type: object
  request.RegisterDebtorPayload:
    properties:
      date_of_birth:
        minLength: 1
        type: string
      full_name:
        maxLength: 255
        minLength: 1
        type: string
      identity_picture_url:
        maxLength: 2048
        minLength: 1
        type: string
      legal_name:
        maxLength: 255
        minLength: 1
        type: string
      nik:
        maxLength: 25
        minLength: 16
        type: string
      place_of_birth:
        maxLength: 255
        minLength: 1
        type: string
      salary:
        type: string
      selfie_picture_url:
        maxLength: 2048
        minLength: 1
        type: string
    required:
    - date_of_birth
    - full_name
    - identity_picture_url
    - legal_name
    - nik
    - place_of_birth
    - salary
    - selfie_picture_url
    type: object
  request.RegisterPayload:
    properties:
      email:
        type: string
      password:
        maxLength: 25
        minLength: 5
        type: string
    required:
    - email
    - password
    type: object
  response.DebtorInstallmentLineResponse:
    properties:
      due_date:
        type: string
      id:
        type: string
      installment_amount:
        type: string
      installment_number:
        type: integer
      payment_date:
        type: string
      status:
        type: string
    type: object
  response.DebtorInstallmentResponse:
    properties:
      debtor_transaction:
        $ref: '#/definitions/response.DebtorTransactionResponse'
      end_date_period:
        type: string
      id:
        type: string
      lines:
        items:
          $ref: '#/definitions/response.DebtorInstallmentLineResponse'
        type: array
      monthly_installment:
        type: string
      paid_installment_count:
        type: integer
      start_date_period:
        type: string
      tenor_duration:
        type: integer
      total_installment_count:
        type: integer
    type: object
  response.DebtorResponse:
    properties:
      date_of_birth:
        type: string
      full_name:
        type: string
      id:
        type: string
      identity_picture_url:
        type: string
      legal_name:
        type: string
      nik:
        type: string
      place_of_birth:
        type: string
      salary:
        type: string
      selfie_picture_url:
        type: string
      user_id:
        type: string
    type: object
  response.DebtorTenorLimitResponse:
    properties:
      current_limit:
        type: string
      tenor_duration:
        type: integer
      tenor_limit_type:
        type: string
      total_limit:
        type: string
    type: object
  response.DebtorTransactionResponse:
    properties:
      admin_fee:
        type: string
      asset_name:
        type: string
      contract_number:
        type: string
      id:
        type: string
      otr:
        type: string
      status:
        type: string
      total_interest:
        type: string
      total_loan:
        type: string
    type: object
  response.ListDebtorInstallmentResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/response.DebtorInstallmentResponse'
        type: array
      total:
        type: integer
    type: object
  response.ListDebtorResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/response.DebtorResponse'
        type: array
      total:
        type: integer
    type: object
  response.LoginResponse:
    properties:
      token:
        type: string
      user:
        $ref: '#/definitions/response.UserResponse'
    type: object
  response.RegisterDebtorResponse:
    properties:
      date_of_birth:
        type: string
      debtor_tenor_limits:
        items:
          $ref: '#/definitions/response.DebtorTenorLimitResponse'
        type: array
      full_name:
        type: string
      identity_picture_url:
        type: string
      legal_name:
        type: string
      nik:
        type: string
      place_of_birth:
        type: string
      salary:
        type: string
      selfie_picture_url:
        type: string
    type: object
  response.RegisterResponse:
    properties:
      email:
        type: string
      role:
        type: string
      user_id:
        type: string
    type: object
  response.UserResponse:
    properties:
      role:
        type: string
      user_id:
        type: string
    type: object
  validations.ErrorField:
    properties:
      field:
        type: string
      tag:
        type: string
      value:
        type: string
    type: object
  validations.ErrorValidation:
    properties:
      fields:
        items:
          $ref: '#/definitions/validations.ErrorField'
        type: array
      message:
        type: string
    type: object
externalDocs:
  description: OpenAPI
  url: https://swagger.io/resources/open-api/
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Debtor API
  version: 1.0.
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: Login to get token
      parameters:
      - description: Body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.LoginPayload'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.LoginResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/validations.ErrorValidation'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/validations.ErrorValidation'
      summary: Login
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Register Debtor User
      parameters:
      - description: Body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.RegisterPayload'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.RegisterResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/validations.ErrorValidation'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/validations.ErrorValidation'
      summary: Register
      tags:
      - auth
  /debtor/detail:
    get:
      consumes:
      - application/json
      description: Detail Debtor of logged in User
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.DebtorResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/validations.ErrorValidation'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/validations.ErrorValidation'
      summary: Detail Debtor
      tags:
      - debtor
  /debtor/installment/list:
    get:
      consumes:
      - application/json
      description: List Debtor Installment of logged in User
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ListDebtorInstallmentResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/validations.ErrorValidation'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/validations.ErrorValidation'
      summary: List Debtor Installment
      tags:
      - debtor
  /debtor/installment/pay:
    post:
      consumes:
      - application/json
      description: Pay Debtor Installment Line of logged in User
      parameters:
      - description: Body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.DebtorPayInstallmentLinePayload'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.DebtorInstallmentLineResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/validations.ErrorValidation'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/validations.ErrorValidation'
      summary: Pay Debtor Installment Line
      tags:
      - debtor
  /debtor/register:
    post:
      consumes:
      - application/json
      description: Register Debtor to get Tenor Limits
      parameters:
      - description: Body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.RegisterDebtorPayload'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.RegisterDebtorResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/validations.ErrorValidation'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/validations.ErrorValidation'
      summary: Register Debtor
      tags:
      - debtor
  /debtor/transaction:
    post:
      consumes:
      - application/json
      description: Create a new Debtor Transaction
      parameters:
      - description: Body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.DebtorTransactionPayload'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.DebtorTransactionResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/validations.ErrorValidation'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/validations.ErrorValidation'
      summary: Create Debtor Transaction
      tags:
      - debtor
  /list/debtor:
    get:
      consumes:
      - application/json
      description: List all registered Debtor
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ListDebtorResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/validations.ErrorValidation'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/validations.ErrorValidation'
      summary: List Debtor
      tags:
      - admin
swagger: "2.0"